+++ Control/DeepSeq.hs
@@ -92,6 +92,7 @@ import Data.Ord ( Down(Down) )
 
 #if MIN_VERSION_base(4,7,0)
 import Data.Proxy ( Proxy(Proxy) )
+import Data.Type.Equality
 #endif
 
 #if MIN_VERSION_base(4,8,0)
@@ -388,6 +389,22 @@ instance NFData Word64   where rnf = rwhnf
 instance NFData (Proxy a) where rnf Proxy = ()
 -- |@since 1.4.3.0
 instance NFData1 Proxy    where liftRnf _ Proxy = ()
+
+-- | @since 1.4.3.0
+instance NFData (a :~: b) where rnf = rwhnf
+-- | @since 1.4.3.0
+instance NFData1 ((:~:) a) where liftRnf _ = rwhnf
+-- | @since 1.4.3.0
+instance NFData2 (:~:) where liftRnf2 _ _ = rwhnf
+#endif
+
+#if MIN_VERSION_base(4,10,0)
+-- | @since 1.4.3.0
+instance NFData (a :~~: b) where rnf = rwhnf
+-- | @since 1.4.3.0
+instance NFData1 ((:~~:) a) where liftRnf _ = rwhnf
+-- | @since 1.4.3.0
+instance NFData2 (:~~:) where liftRnf2 _ _ = rwhnf
 #endif
 
 #if MIN_VERSION_base(4,8,0)